Update request.rb
I have debugged the code and found that httparty response.parsed_response is returning a json string object. so wee need to parse it as json object. raise_error(response['error'], res.code) if response.key?('error') this line is producing error (NoMethodError for key? on String). So we need to update the above code.